move click tests to dedicated file
